Added Svelte to the dependancies (#141)
* Added Svelte to the dependancies 

When I cloned this project and installed the dependencies, I then added the files from the `README.MD` file, but I was unable to get this project to work without installing Svelte as a normal dependency. I updated the the `README.MD`  with instructions on how to add Svelte as a dependency.

I do not have Svelte installed globally on the instance that I'm using it. So maybe that is why I needed it as a dependency to run.

* Moved Svelte to Dev Dependency

Turns out it works just fine if you add Svelte as a dev dependency instead of a regular dependency.